+++ testNet60.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-dotnet test -f net6.0